Add new matches to event
