+++ src/routes/events/[id]/add/+page.svelte
@@ -1,34 +1,62 @@
 <script lang="ts">
 	import GamePicker from '$lib/components/GamePicker.svelte';
-	import { data, type Game, type Map } from '$lib/data';
+	import { library, type Game, type GameMap, type GameMode } from '$lib/library';
 	import MapPicker from '$lib/components/MapPicker.svelte';
 	import ModePicker from '$lib/components/ModePicker.svelte';
 	import Grouping from '$lib/components/layout/Grouping.svelte';
+	import { db } from '$lib/db';
+	import { goto } from '$app/navigation';
+
+	/** @type {import('./$types').PageData} */
+	export let data: any;
 
 	let selectedGame: Game;
+	let selectedMode: GameMode;
+	let selectedMap: GameMap;
 
 	function setSelectedGames(event: CustomEvent<Game[]>) {
 		selectedGame = event.detail[0];
+	}
+
+	function setSelectedModes(event: CustomEvent<GameMode[]>) {
+		selectedMode = event.detail[0];
+	}
+
+	function setSelectedMaps(event: CustomEvent<GameMap[]>) {
+		selectedMap = event.detail[0];
+	}
+
+	async function confirm() {
+		await db.createMatch(data.params.id, {
+			gameId: selectedGame.id,
+			modeId: selectedMode.id,
+			mapId: selectedMap.id
+		});
+		await goto(`/events/${data.params.id}`);
 	}
 </script>
 
 <div class="container">
 	<Grouping>
 		<span slot="title">Game</span>
-		<GamePicker games={data.games} multi={false} on:selectionChange={setSelectedGames} />
+		<GamePicker games={library.games} multi={false} on:selectionChange={setSelectedGames} />
 	</Grouping>
 	<Grouping>
 		<span slot="title">Game Mode</span>
-		<ModePicker modes={selectedGame?.modes} />
+		<ModePicker modes={selectedGame?.modes} multi={false} on:selectionChange={setSelectedModes} />
 	</Grouping>
 	<Grouping>
 		<span slot="title">Map</span>
-		<MapPicker maps={selectedGame?.maps} multi={false} />
+		<MapPicker maps={selectedGame?.maps} multi={false} on:selectionChange={setSelectedMaps} />
 	</Grouping>
 </div>
 <div class="button-region">
 	<a class="pure-button" href="./">&times;&nbsp;Cancel</a>
-	<button class="pure-button button-primary">&check;&nbsp;Confirm</button>
+	<button
+		class="pure-button button-primary"
+		on:click={confirm}
+		disabled={!selectedGame || !selectedMap}>&check;&nbsp;Confirm</button
+	>
 </div>
 
 <style>
