+++ src/routes/events/[id]/add/+page.ts
@@ -1,0 +1,6 @@
+/** @type {import('./$types').PageLoad} */
+export function load({ params }: any) {
+    return {
+       params: params
+    };
+}
